requirements.txt
fastapi>=0.105.0
uvicorn>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.23
qdrant-client>=1.12.0
sentence-transformers>=3.0.0
fastembed>=0.3.1
cohere>=4.0
openai>=1.3.7
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
python-dotenv>=1.0.0
pytest>=7.4.3
httpx>=0.25.2
markdown>=3.5.1
beautifulsoup4>=4.12.2
trafilatura>=1.6.0
requests>=2.31.0
nest_asyncio>=1.6.0
jose>=1.0.0
email-validator>=2.0.0
better-auth==0.0.1b11